Add more factory functions for BasicAdapter
